% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{bart}
\alias{bart}
\title{Combined Semi-Parametric BART}
\usage{
bart(
  y,
  x,
  sparse = FALSE,
  ntrees = 50,
  node_min_size = 5,
  alpha = 0.95,
  beta = 2,
  nu = 3,
  lambda = 0.1,
  mu_mu = 0,
  sigma2 = 1,
  sigma2_mu = 1,
  nburn = 1000,
  npost = 1000,
  nthin = 1
)
}
\arguments{
\item{y}{x}

\item{x}{x}

\item{sparse}{x}

\item{ntrees}{x}

\item{node_min_size}{x}

\item{alpha}{x}

\item{beta}{x}

\item{nu}{x}

\item{lambda}{x}

\item{mu_mu}{x}

\item{sigma2}{x}

\item{sigma2_mu}{x}

\item{nburn}{x}

\item{npost}{x}

\item{nthin}{x}
}
\value{
x
}
\description{
Combined Semi-Parametric BART
}
\examples{
#

}
