% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cspbart.R, R/predict_tree.R
\name{cspbart}
\alias{cspbart}
\alias{predict.cspbart}
\title{Combined Semi-Parametric BART}
\usage{
cspbart(
  formula,
  x1,
  x2,
  sparse = FALSE,
  ntrees = 50,
  node_min_size = 5,
  alpha = 0.95,
  beta = 2,
  nu = 3,
  lambda = 0.1,
  mu_mu = 0,
  sigma2 = 1,
  sigma2_mu = 1,
  nburn = 1000,
  npost = 1000,
  nthin = 1
)

\method{predict}{cspbart}(object,
        newdata_x1,
        newdata_x2,
        type = c('all', 'median', 'mean'),
        ...)
}
\arguments{
\item{formula}{x}

\item{x1}{x}

\item{x2}{x}

\item{sparse}{x}

\item{ntrees}{x}

\item{node_min_size}{x}

\item{alpha}{x}

\item{beta}{x}

\item{nu}{x}

\item{lambda}{x}

\item{mu_mu}{x}

\item{sigma2}{x}

\item{sigma2_mu}{x}

\item{nburn}{x}

\item{npost}{x}

\item{nthin}{x}

\item{object}{x}

\item{newdata_x1}{x}

\item{newdata_x2}{x}

\item{type}{x}

\item{...}{Catches unused arguments.}
}
\value{
x
}
\description{
Combined Semi-Parametric BART
}
\examples{
#

}
