% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_tree.R
\name{var_used_trees}
\alias{var_used_trees}
\title{Get covariates used as splits per iteration}
\usage{
var_used_trees(object, raw = FALSE)
}
\arguments{
\item{object}{x}

\item{raw}{x}
}
\value{
x
}
\description{
Get covariates used as splits per iteration
}
